Ensure that specific header fields are present only once
The following usage of API is creating RFC not compliant emails:
 QMailMessage m;
 m.setFrom(QMailAddress("bla@bla.org"));
 m.appendHeaderField("From:", "bla@bla.org");
This construction may happen for instance if the appendHeaderField() call is done
in a for loop copying all headers from another email.

Several header fields from RFC2822, 3.6, are mentioned to be present at
maximum only once. A filter based on field name is set to change the
append call into an update call for these specific fields.

Change-Id: Iece09438b3252e2a6888b7a6aafb3c9cc280558c
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>