Exclude parts containing signature data from attachment list
The signature data of PGP or S/MIME standards requires to be checked
on the undecoded data of another part of the email. This other part
is available internally through the undecoded() method, but is not
available outside QMF. There is thus no point to propose the signature
data as an attachment that may become visible to the user.

This patch remove the signature data of these two protocols from the
list of attachment, as it was done for the calendar data for instance.

It also updates and extends the tests of the findAttachmentLocations()
method.

Since imap plugin is using findAttachmentLocations() to decide to retrieve
parts or full message, a modification was required to ensure that
signature data are properly parsed in case of full message download.

Change-Id: If0da27ce06d78b15a44c40bfb409f84635a3a8f6
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>