Don't download twice part content in case of signed message with IMAP protocol
There is an issue with the current implementation that downloads twice the
part content in case of subparts and signed mails. For instance a mail with
this layout:
 1: multipart/signed
 1.1: text/plain
 1.2: image/png
 2: application/pgp-signature
will trigger in mode RetrievalAction::Auto the following requests:
 1-HEADER  // as a signed part
 1-BODY    // as a signed part
 1.1-BODY  // as a text part
 2-BODY    // as a non attachment part
which will result in text/plain part being downloaded twice and appended twice
in a text file. To avoid this, this patch don't append to the list of requests
any requests that are within a signed part.

Change-Id: I654d020acf2a6a1fe26ae53ad667a84e7ea90c2e
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>