Ensure that setting empty preview results in non-null value
QMailStore specifies a NOT NULL constraint for the preview
field of the mailmessages table.  If a client attempts to
clear the preview for the message by passing in a
default-constructed QString(), the subsequent database operation
will fail due to the violated constraint.

This commit ensures that the preview field is handled similarly
to other fields (like rfcId) such that a non-null (but still
empty) string value is used in that case.

Change-Id: Ia527adea1f68de4d59b8837c3f8ce288d454a7d7
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>