Decrease exists count on EXPUNGE
According to the IMAP spec (RFC3501), EXPUNGE messages indicate a
decrease by one in the number of messages on the server, and the server
doesn't have to notify the client of this decrease with a separate
EXISTS message (from Section 7.4.1):

> The EXPUNGE response also decrements the number of messages in the
> mailbox; it is not necessary to send an EXISTS response with the new
> value.

Moreover an EXISTS message cannot be used on its own to indicate that
the number of messages has reduced (from Section 5.2):

> Special rules exist for server notification of a client about the
> removal of messages to prevent synchronization errors; see the
> description of the EXPUNGE response for more detail.  In particular,
> it is NOT permitted to send an EXISTS response that would reduce the
> number of messages in the mailbox; only the EXPUNGE response can do
> this.

Currently when messageserver receives an EXPUNGE message it doesn't
decrease the number of messages recorded in the 'exists' count. This can
cause the number recorded by the client and server to become
desynchronised, because some server implementations don't send an EXISTS
directly after an EXPUNGE (which is consistent with the spec as
summarised above).

This can cause problems when using IDLE connections, because an EXISTS
message sent on an IDLE connection will only trigger a new mail event if
the number of messages indicated by the server is greater than the
number recorded by the client. Therefore if the server sends an EXPUNGE
message that isn't followed by an EXISTS message, the client will record
a higher 'exists' value than the number on the server. When a new
message arrives at the server, the server will notify the client with an
EXISTS message, which the client will then erroneously ignore.

This change fixes the issue by decreasing the 'exists' variable by one
every time an EXPUNGE message is received.

Change-Id: I00495bb83776f2320754d58ae0ec4b091f90c182
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>