Fix QMailMessageListModel getting duplicated content
If code using this model first calls setKey(), the model will be marked
as uninitialized and it emits reset signals.
If the next call is setLimit(), the code here used populate the content
(albeit it left the mode still uninitialized so the next access populated
the content again anyway).

If nothing used the model above worked fine, even if a bit suboptimal,
but now if e.g. a view acted upon [begin/end][Remove/Insert]Rows
signals by accessing the model, the initialization population was done
in the middle of setLimit() still doing its own population.

Change-Id: Id3c028ae477459e1d7d0fe94ee62aab08a29318b
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: David Llewellyn-Jones <david.llewellyn-jones@jolla.com>
Reviewed-by: Joona Petrell <joona.petrell@jollamobile.com>