Add mock NetworkStatusMonitor class to SMTP plugin
The SMTP plugin previously used QNetworkConfigurationManager to
determine network connectivity state.  This class has been removed
in Qt6, and will in future be replaced by QNetworkInformation.

This commit replaces the use of QNetworkConfigurationManager with
a mock NetworkStatusMonitor class, which will in future be replaced
by QNetworkInformation once that class is available and released.

Change-Id: Iff415dc974bf8d782719ad7a77c653adb81e7f5d
Reviewed-by: Christopher Adams <chris.adams@jolla.com>
Reviewed-by: David Llewellyn-Jones <david.llewellyn-jones@jolla.com>