Switch to DBus for IPC
There are two IPC mechanisms in QMF:
- one between the messageserver and clients
  to perform actions like fetch emails,
  search, etc.
- one between every mailstore instances
  to notify other instances of changes
  in the mail database.

QCop being discontinued, this patch switchs
to use DBus as an IPC transport layer.

The first mechanism is implemented by registering
an object on the session bus (/messageserver)
providing a service (org.qt.messageserver).

The second mechanism is implemented by
registering each client as an object to the
session bus (/mailstore/client) for the
interface org.qt.mailstore. Other clients
can then listen to updated() signal on this
interface.

DBus IPC requires to implement operator<<()
and operator>>() with QDBusArgument for the
objects that need to be transported. In all
cases, each objects are serialised up to
unitary elemental elements like int, byte
arrays, etc. Except for the QMailKeyArgument
that uses first a serialisation into a
byte array before being sent. This is due to
the recursive declaration of QMailKeyArgument
with the subKeys members, not working well
with QDBusArgument serialisation.

Change-Id: I8f10b8b00f73d4b75dad778b591159a765e9fb89
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>