Delete deprecated message count signals in IPC
The newCountChanged signal is not used anymore
in the message server. Remove the client code
exposing it.

Remove also client method acknowledgeNewMessages()
since it's linked to newCountChanged signal to
reset the counters.

Remove the internal QMailMessageServerPrivate::initialize()
since it's not called from anywhere and the server
has no slot for this signal.

Change-Id: Id5a67c22eae2c7190c3e02ed0681eab7037f1fcc
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>