Listen to activityChanged in action observer
There is a race condition in the way the observer
is listening on activity change for the actions it
is supervising. Indeed, the servicehandler is
emitting actionStarted() signal, and then
activityChanged(). The observer may receive these
signals asynchronously. But with the current
implementation, the observer starts listening at
activityChanged() after it has received the
actionStarted() one. The race condition appears
then if the servicehandler emits the activityChanged
before the observer received the actionStarted()
one.

This patch should correct this race condition by
making the observer always listen to activityChanged()
and deal with it if the activity corresponds to an
action it is supervising.

Change-Id: I339be274152e648dda50e8bb4667a8fdd70e4cba
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>