Share message server with ActionInfo
The ActionObserver is responsible to spawn
ActionInfo objects when actionStarted() is
received from IPC. Then, the newly created
ActionInfo listens as normal Action for update
signals over IPC.

But, depending when the ActionInfo creation
is actually done, it may have missed IPC update
signals. To avoid this, the observer is sharing
its QMailMessageServer, and the newly created
ActionInfo can connect to signals that are still
pending in the message server.

Change-Id: I152fdcbeee7a902a0c54b58a28476cdfb31d4dbc
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>