Don't readd received IPC signals in the event loop
With QCop IPC implementation, received
signals where put in a message queue to
be processed in the next event loop
iteration. This is not needed anymore
with the DBus implementation, since
the receiving process lives in a different
thread and already submit received
signals in the event loop of the main
thread.

Change-Id: I4ebfb2b03f7ab21a27642d380b5dc2fe5c281791
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>