Set replied and forwarded flags on message update from server
The replied flag (\Answered in IMAP protocol) is
properly set for a new mail. But it was not set
on a qresync action (fetching mail updates).

The forwarded flag ($Forwarded as an extension
in IMAP protocol), was detected in the received flags
but was not set or updated in created QMailMessages.

Change-Id: I78eb16650dc706e08fee1c6aab68d601159a52e4
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>