Fix build on Qt6
Regression on the D-Bus ipc which introduced
e.g. Q_DECLARE_USER_METATYPE(QMailMessagePart) on qmailmessage.h.

On Qt6 the qRegisterMetaType() ends up instantiating
template<QDebug> QMailMessagePartPrivate::serialize(type)
which then fails to build due to not having << operator for
combination of QDebug and QMailMessagePartContainer::Location.
Compiler ends up spewing pages and pages of errors and warnings.

Not entirely following what's causing the difference to Qt5, and
how QDebug serialize() works with one class but not on other, but
suppose this explicit operator should be clean enough and avoiding
side-effects on the actually running code.

Change-Id: I4939e89d625ac06c682adbb5e31803e1ef019fd1
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>