Reset parts when setting a body
Currently, setBody() is called to set content to an
empty part. Everywhere in the code, this method is
called like an initializer, or behind a test checking
that the part is indeed MultipartNone.

But setBody() being in the public API, nothing
restrain the user to call it on a part that
already contains something, including several
subparts.

Ensure that subparts are removed when setting
a body with content. Similar to what is done in
addAttachmentsToMultipart(), unset the attachment
flag if the part we set a body to is actually a
message itself.

Change-Id: I245c5543cd85afed1176987e190900ff930efdbb
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>