Add a fetchCapabilities() method in SmtpClient
With this new method, a SmtpService can inquire the
capabilities without having to fake a message transmission
without any message. The capability fetching is
done with a dedicated client, so actions sent to the
service can be honoured while fetching.

Fetching capabilities does not require to log in, thus
it can be run any time, even with unset credentials.

Capabilities are reflected in new entries in the account.
They are fetched when the service is created for a given
account.

Change-Id: Ibcefc2f92fdaa5ec3f36f937db8d8b278b3a4f6a
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>