Move LongStream to server library, streamline the implementation
The qmfclient was using this just to get storage info, which is
somewhat out of scope for the class api anyway. And also the
"errorMessage" meaning out-of-space error was bad API.

Got rid of OS specific storage calculation in favor of QStorageInfo
and streamlined the error string method.

To me the implementation feels like trying a bit too much to check the
available space, but I'll not touch that now.

Change-Id: Ice4256c0cdf3fa3dc52c1354ab07a433862e66ae

Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>