Remove win32_ce traces and some unnecessary platform specific bits
Don't think we'll need win32ce / windows mobile support by now.
Not sure how beneficial it was in the first place if it lacked
database support.

Besides that, the whole ifdeffing in qmailnamespace.cpp seemed broken
here, block starting outside methods and ending somewhere in a middle
of a method.

pop3client unistd.h not needed anymore, was adder for ::sync()
but that is already gone.

Change-Id: I530028f2fd7d476df0ad11ea259dfb843c627352
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>