Reduce status variable number in ImapClient::monitor()
Makes the idle state to rely on _monitored and
_waitingForIdleFolderIds variables only.

_monitored is the set of IdleProtocol listening
to changes in a folder. _waitingForIdleFolderIds
is a subset of _monitored, listing folder where
the protocol was started, but has not reached yet
the idling status.

_idlesEstablished can be replaced by a combination
of the two, being true when and only when some
folders are monitored and all of them are in the
idling state.

Old variable _waitingForIdle is not needed anymore
if monitor() function is modified to ensure that
the list of folders given as argument is properly
transformed into monitored folders, and only these
folders as argument. Calling monitor on the same
list is a no operation, making _waitingForIdle
completely obsolete.

A new closeIdleConnections() function is added
to the public API of ImapClient, since there was
previously no mean to close idle connections,
except by killing the client itself. This is
achieved internally by closing non desired
idle protocols in monitor() before checking if
the settings are push enabled.

Change-Id: I8c62b6d5ff7150ee2c90d7dcfec3bd44f66cf7b6
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>