Rework ImapService::onAccountsUpdated()
Avoid the case of calling disable() in the
callback for any other case than when the account
is disabled.

In particular, don't call disable() + enabled()
to apply a change. The only account change that
requires to stop (and restart) a running client
is a change in the idle folder list. In that case,
there is no need to kill and restart the client,
just restarting the idle function is enough.

Change-Id: Ica39c1c09bb5b156670a0f1301bc6dc9703a7451
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>