Report client error even when establishing idle connections
The errorOccurred signal is only emitted for client
session, and not for idle connections (there are
reporting via the client itself, which emits
restartPushEmail signal for idle errors).

It is thus necessary to report any error the client
is emitting. If the client session was launched to
establish idle connections and these sessions are not
running, then one should retry to reestablish the
idle connections later.

Without the patch, in case of a non responding server
(connection to a captive portal without network
granted yet), the client error was not reported when
the connection could not be established. It was
accepted wrongly as succeeded when a new attempt to establish
idle connections was made and Source::retrievalCompleted()
was emitted when queueing the idle folder for scan.
But with the retry timeout becoming larger and larger,
the strategy expiry timeout finally kicked in and the
service was fully restarted.

Change-Id: I206e77f221d2a7c4bdfcee1952dbe1127431e6d6
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>