Make ImapService the only place to delay push mail restart
The IdleProtocol was delaying its signal emission
of openRequest in case of an error. Then the client
emitted restartPushEmail and the service was immediately
calling initiatePushEmail(). This is the case when
the error happened in the idle protocol.

But the service is already having a delayed machinery
to restart the idle session in case an error happens
in the main client session.

This patch is removing the delay timer in IdleProtocol
so any error happening there is immediately reported to
the service, which decides to delay the call to
initiatePushEmail(). This requires some safe guards
to be added in the service to ensure that the delay
is applied only once even when there are several
push folders emitting errors.

Change-Id: I2b49c0445640d58bb391088b602f152cfe45eb49
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>