Continue login when there is no valid push folders
Currently, login is deferred until all
configured push folders have been established
(or failed), when push email is enabled.

This does not work for a newly created account
where there is no folder registered.

This patch ensures that the code continues
to the login when push email is enabled but
when there is no valid configured folders.

Change-Id: I397d4d2616c3c6cebc9a3f4dcc03f4b5e3564ebe
Reviewed-by: Guido Berhoerster <guido+qt@berhoerster.name>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>