Declare tests in the build system
Tests can be run with `ctest`. The `--verbose`
option makes the output of each test visible.

Fix also minor failures:
- make qlogsystem test locale independent,
- make qmailstore test properly listening to
  the messagePropertyUpdated() signal.

IMAP, POP and SMTP tests are declared as always
failing for the moment, since they try to
connect to a mail server (but they are still
useful for debug purposes).

Change-Id: I7ae548110a8eeb37bd07b031a8433d38ea53529e
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>