Adjust and fix protocolRequest API - minor API break
QMailProtocolAction::protocolRequest() is specified
"The request may have associated \a data, in a protocol-specific form",
which should mean that invalid QVariant() is allowed to indicate no data.

This worked earlier but the D-Bus IPC broke it due to D-Bus wire protocol
not allowing invalid variant, and QDBus was refusing to handle such:

 'QDBusConnection: error: could not send message to service
  "org.qt.messageserver" path "/messageserver" interface
  "org.qt.messageserver" member "protocolRequest": Marshalling failed:
  Variant containing QVariant::Invalid passed in arguments'

Side-stepped the problem by switching type to QVariantMap, which
is simple to serialize empty and named parameters can be passed now
without extra hassle. It's a small API break but the protocol requests
are quite an advanced feature, and not used with the QMailMessageSource
instances in this repository.

Change-Id: I0a44026938d6e77bc7a89640cffe6efa38cc5616
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>