Avoid custom QPrivatelyNoncopyable on QMailServiceAction - ABI break
Just use the common Q_Q and Q_DECLARE_PRIVATE. Not seeing a point in the
custom code that makes it harder to follow. The QPrivateImplementation
talks about QSharedDataPointer. I'm not entirely following what's
really the problem but in this direction such is not even used.

Small ABI break: apps using QMailServiceAction likely need to be
recompiled due to instance size change.

Change-Id: I0fe87f277ff1c2383331dfe9e9e29377ff608479
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>