New connect() api in QMailServiceAction - small api break
New api better checks the types and we have complex enough ones here.

Besides that, the api had overloaded names for
QMailSearchAction::remainingMessagesCount() and ::messagesCount(),
having same name for both signal and getter. While we could cope with
that with some extra code, it's really just bad API so changed
the signal names to indicate it's about change.

Also the totalProgressChanged() signalling seemed broken.
The signal got connected to handler in private class, which
emitted a similar signal, but that wasn't connected to the main class.
Added such now in QMailActionInfo() ctor.

Change-Id: If0264d8de65509cd1be9b5a5d994c57ab9e9f0d5
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>