Fix account added twice in list model
If an account is added, while an account list model is still
uninitialised, the account was added twice. Correct it by
returning when d->initialize() is called. Also exclude current
ids when matching new accounts. Also rename init member and
initialize() to populated and populate() for clearer meaning.

Additional clean-up in the class, polishing spacing and
brackets alignments.

Change-Id: I1e5e3b8dae15d0b86bfbf26b48085d1d9b599707
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>