Handle API TODO item from ImapTransport
There was a note for making some methods virtual in the base class
so the 'imap' prefix could be avoided. Did that first but then reviewing
the situation it felt better to keep the prefix on imap side and on
QMailTransport exposing socket and stream means that overriden methods
can be too easily side-stepped.

Originally the TODO was referring to a) imapReadLine() but
base has an optional parameter which would have needed to be implemented,
even if not needed, b) imapCanReadLine() where matching base
is const while this one needs non-const access to consume compressed
data, and c) imapWrite() which doesn't have a matching parent method.

Later added bytesAvailable() / imapBytesAvailable() and
readAll() / imapReadAll() could have been made virtual but it would have
been also confusing if some reads were having special methods and some
not.

For suspicious earlier behavior, ImapProtocol was having one place where
it was calling bytesAvailable() and readAll() instead of the
imap variants, avoiding the potential decompression. Coming from
commit 37164dc1b which did even introduce the new methods in
ImapTransport but no usage for them.

Change-Id: I3eec5034b8393b26f288bf0a158c8c1b1d520445
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>