Refactor ServiceHandler and avoid unnecessary data serialization
The serialization of request data is not only ugly but seemed
problematic when trying to adjust the IPC serialization.
Thus made the code use proper classes for storing the request
related data.

Was initially aiming at moving also the request code to some
'virtual bool execute()' type of method, but that wasn't trivial
due to requests needing some private api to change the ServiceHandler
state. But maybe this is a nice first phase anyway, limiting the
changes done at once.

While at it, did a bunch of other cleanups, simplifications,
small fixes and comments for future work:
- renamed reportFailures as reportPastFailures() to better indicate
what it's about.
- renamed Request::description as requestType which should better
describe the member.
- Moved internal ExpirySeconds from header to cpp
- simplified markFailedMessage() to mark the failed message
with the check for such need on calling place
- minor functional change: retrieveNewMessages() wasn't doing the usual
usesConcurrentActions() thing properly. Now passing the action id
as rest.

Change-Id: I8261c50492d7d496df777e69dbbc22f5485a87fa
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>