Clean up Smtp client not to use fragile iterators and qmflist
Having an iterator member variable for currently sending email
and another iterator for handling its rcpt list was fragily and
unconventional. Just storing now the needed data separately and
getting rid of the email from the list of to be send after it's
done.

The rcpt iterator was also badly having a generic name 'it' etc.
Bunch of other cleanups while at it and avoiding QmfList
when there doesn't seem to be any need.

Change-Id: I98b7829ddea36ec30e50ab72cf3f2599e6e60362
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>