Deprecate Fwd classes in favor of just introducing enums in real classes
The Fwd classes were inherited by the actual classes, allowing to #include
only the fwd and not the bigger real class header. Such a compilation time
micro optimization shouldn't really be worth complicating the API.
Thus just moved the enum content to actual classes.

For some backwards compatibility made the Fwd classes typedeffed to the
real ones.

Crypto side didn't even have a proper class, only the Fwd and suppose
all the usage of the code just did include anyway for the whole crypto api.

The QMailFolderFwd::PredefinedFolderId was having a cyclic dependency
when moved to QMailFolder: QMailFolderId ctor variant was depending on
it and QMailFolder was depending on the Id. Moved the predefined id to
the Id class which feels like a better home anyway.

Change-Id: I9a9aa9c407ae90cc11b37179fe417e3a002ce94b
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: <matti.viljanen@kapsi.fi>