Avoid (explicit) QmfList outside the library
Paving the way to get rid of the QmfList which was added sort of
stopgap on Qt6 migration because of the QList changes.
- Imap plugin using for() loop which should work for any container type
returned by the methods.
- tst_qmailstore unit test just getting the list of addresses after we
are sure those won't change. A few more lines but avoids the custom
type.
- On previous changed also indexing fully 0 based to keep it more robust.
"Hi 0" type of subjects shouldn't really matter on tests.

Change-Id: I7a74dbe7654d5f3dd80435d5bf9af110b875b03a
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>