Remove QMail::messageServerPath() and messageSettingsPath()
Not used in the library itself and the implementations don't make too
much sense: the paths are using QCoreApplication::applicationDirPath()
referring the random client app binary path so there's no much idea
where the qmf content really is.

There is some use for the server path in an example project, perhaps
even proving how little it makes sense as it tries to launch a specific
hard-coded binary from the directory.

If something, the API should have a method for requesting launching the
messageserver.

Change-Id: I9b4b245e5c17f2e49fe3b9ed8b30cb175c5e8e45
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>