Remove non-virtual private method override
There are quite a lot of these indicativeSize() implementations
on public and private side. Making one easily confused.

In this case there's basic implementation in
QMailMessagePartContainerPrivate base class of QMailMessagePrivate,
and then a version doing +1 for header in both non-virtual
QMailMessagePrivate method and the public QMailMessage method.

I'm guessing this was maybe implementation side confusion whether
to do the addition in public method or the private one, ending up
both and some vagueness which private method gets really called
(should be the QMailMessagePrivate and not the base class as
partContainerImpl() returns a qmailmessageprivate pointer, heh).

IndicativeSize not having a precise meaning it probably doesn't matter
too much but good to get rid of a bit of this mess.

Change-Id: I8f8ce6ee459fc98ce56e321a71ae849a65055f5c
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>