Adjust messageserver naming versions
On Qt6 it's strange to have messageserver5. Could have bumped the
binary there to '6' but maybe we don't really need to highlight the
qt version that much. The '5's were added to allow parallel
installability but that really matters more on libraries etc rather
than the server. We don't want parallel installed daemons fighting who
gets autostarted first.

Also on credential side it shouldn't matter which was the Qt version.
Switching to newer version should still be about the same service.
In practice don't think the names matter too much, and auth success is
ignored so think it should be ok just changing the naming.

Removed the '5's from examples too. No need there.

Left the .pro file binary as it was as that side now mostly serves the
Qt5 builds where we wouldn't benefit much from changing the binary name.

Change-Id: Ie4a843c944caa73b080e3610baee4842500f029b
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>