Fix some more compiler warnings
Disclaimer: didn't actually test the example as those are not in good
shape, but the migration to non-deprecated API should be straightforward
enough: mostly question() having default yes/no, and warning() having ok
button is enough to skip explicit button declarations.

For minor change not having now "no" as option when there is
no question in verifyAccount(). checkMailConflict() removed as unused.

Change-Id: I15067e1222264a685ba8c83c50e6ae06d4c51c38
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>
Reviewed-by: <matti.viljanen@kapsi.fi>