Avoid pointless double indirection pointer on writing rfc2822
Not following why such existed here. Simplified by just using a plain
pointer.

Also added a note to the public method that QDataStream is used for
writing plain text. Wondered for a while if the data versioning etc
would matter. On a glance a bit strange API but guess it keeps control
of the encoding etc.

Change-Id: I36349123a54893d85a1f0185182fc33300c3a328
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>