Rename QMailMessageMetadata::content as contentCategory
The related functionality was confusing when setContent() didn't
actually set any content but rather the type of it. Also there are
other setContent() methods around that actually set the content.

Then again calling this contentType() would have collided further
in the inheritance tree by method returning a QMailMessageContentType
which is referring to header fields. Thus switched this to a third
distinct name.

Old setter and getter names provided for backwards compatibility
which should work at least for the very basic cases of setting and
comparing type.

Change-Id: I4c54548f3c2c85096fde64b3221e233dd3feefa3
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>