Mark serialize() methods \internal
Qdoc is having trouble with the templates and was unable to find
content for any of these, instead just complaining. Besides that we
don't really even want to promote this functionality and hopefully
would get rid of it.

Also QMailKeyArgument::ValueList::operator==() documentation didn't
work. Just removed that as minor detail.

Change-Id: I1956c19cf53ed7ef93478d629755b39fbfdbd4bd
Reviewed-by: <matti.viljanen@kapsi.fi>
Reviewed-by: Damien Caliste <dcaliste@free.fr>
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>